Move homepage contents to separate component
