+++ src/comps/home/UnauthenticatedLanding.js
@@ -1,0 +1,9 @@
+import GoogleLoginButton from "../auth/GoogleLoginButton";
+
+const UnauthenticatedLanding = () => {
+  return (
+    <GoogleLoginButton />
+  )
+}
+
+export default UnauthenticatedLanding;
